reorganize firebase deployment
