+++ packages/cli/src/commands/deploy/service.ts
@@ -4,8 +4,7 @@ import { prompt } from 'inquirer'
 import { IDefinition, IService } from '@mesg/api/lib/service'
 import { compile } from '../../utils/service'
 import { loginFromCredential } from '../../utils/login'
-import firebase from '../../utils/firebase'
-import { EventEmitter } from 'events'
+import firebase, { deploy } from '../../utils/firebase'
 
 const ipfsClient = require('ipfs-http-client')
 
@@ -47,11 +46,11 @@ export default class Service extends Command {
         title: 'Creating service',
         task: async (ctx, task) => {
           const title = task.title
-          const logs = await this.deploy('service', definition, user.uid)
+          const logs = await deploy('service', definition, user.uid)
           logs.on('data', x => {
             task.title = `${title} (${x.message})`
           })
-          service = await this.toPromise(logs)
+          service = await logs.promise()
         }
       }
     ])
@@ -61,18 +60,18 @@ export default class Service extends Command {
     this.log(`Service deployed with the hash ${service.hash}`)
     this.log('')
 
-    const { deploy } = await prompt({ type: 'confirm', name: 'deploy', message: 'Would you like to start this service?' })
+    const { start } = await prompt({ type: 'confirm', name: 'start', message: 'Would you like to start this service?' })
 
-    if (deploy) {
+    if (start) {
       const env = await this.getEnv(definition.configuration.env || [])
 
-      const logs = await this.deploy('runner', {
+      const logs = await deploy('runner', {
         serviceHash: service.hash,
         env
       }, user.uid)
       logs.on('data', this.log)
 
-      const runner = await this.toPromise(logs)
+      const runner = await logs.promise()
       this.log('')
       this.log(`Service started with the runner hash ${runner.hash}`)
     }
@@ -89,50 +88,5 @@ export default class Service extends Command {
       type: 'input'
     }))) as { [key: string]: string }
     return Object.keys(res).map(x => `${x}=${res[x]}`)
-  }
-
-  private async deploy(type: 'service' | 'runner', definition: any, uid: string): Promise<EventEmitter> {
-    const res = await firebase.firestore().collection('deployments').add({ type, uid, definition })
-    const eventEmitter = new EventEmitter()
-
-    let unsubscribeLogs: () => void
-    let unsubscribeDeployment: () => void
-
-    const clearListeners = () => {
-      if (unsubscribeDeployment) unsubscribeDeployment()
-      if (unsubscribeLogs) unsubscribeLogs()
-    }
-
-    unsubscribeLogs = res.collection('logs').onSnapshot(
-      snapshots => {
-        for (const change of snapshots.docChanges()) {
-          if (change.type !== 'added') continue
-          const data = change.doc.data()
-          if (data.level === 'error') return eventEmitter.emit('error', new Error(data.message))
-          eventEmitter.emit('data', data)
-        }
-      },
-      error => eventEmitter.emit('error', error)
-    )
-
-    unsubscribeDeployment = res.onSnapshot(
-      async snapshot => {
-        const data = snapshot.data()
-        if (!data.resourceRef) return
-        const resource = await data.resourceRef.get()
-        eventEmitter.emit('end', resource.data().definition)
-      },
-      error => eventEmitter.emit('error', error)
-    )
-    eventEmitter.on('error', clearListeners)
-    eventEmitter.on('end', clearListeners)
-    return eventEmitter
-  }
-
-  private async toPromise(eventEmitter: EventEmitter): Promise<any> {
-    return new Promise((resolve, reject) => {
-      eventEmitter.on('end', resolve)
-      eventEmitter.on('error', reject)
-    })
   }
 }
